Stop double-quoting Mesosphere repo URL
Puppet lint will fail when the URL is quoted because there are no variables
inside which need expanding. Double quotes are only really acceptable when
there are. Fix by changing this to single quotes, which also matches other
apt::repository::url references in this repository, and indeed this file.

Fixes the lint tests which fixes the Travis build.